refactoring and componetizing widgets
